CHANGE: Move image reader from file folder to image folder @zhouxiaoji
