+++ agentuniverse/agent/action/knowledge/reader/image/image_reader.yaml
@@ -1,6 +1,13 @@
-name: 'default_image_reader'
-description: 'default image reader by easyocr'
+name: "default_image_reader"
+description: "Default image reader that supports OCR and image description generation"
 metadata:
   type: 'READER'
   module: 'agentuniverse.agent.action.knowledge.reader.image.image_reader'
   class: 'ImageReader'
+
+default_symbol: true
+languages:
+  - "en"
+  - "ch_sim"
+use_gpu: false
+image_model: "microsoft/git-base-coco"
