+++ tests/test_agentuniverse/unit/agent/action/knowledge/reader/image/test_images/test.jpg
